requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
transformers==4.41.0
accelerate==0.30.0
# bitsandbytes==0.43.1
langchain==0.2.0
langchain-community==0.2.0
chromadb==0.5.0
sentence-transformers==3.0.0
pypdf==4.2.0
pillow==10.3.0
opencv-python-headless==4.9.0.80
paddlepaddle==2.6.2
paddleocr==2.7.3
groq==0.9.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
gradio==4.36.0